LayoutTests/platform/ios-simulator/webgl/2.0.0/conformance/context/context-hidden-alpha-expected.txt

CONSOLE MESSAGE: line 154: FAIL gl.getParameter(gl.ALPHA_BITS) should be 0. Was 8.
CONSOLE MESSAGE: line 138: FAIL Should blend as if alpha is 1.0.
at (150, 75) expected: 64,64,64,255 was 0,0,0,0
CONSOLE MESSAGE: line 138: FAIL dataURL_post should be data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAASwAAACWCAYAAABkW7XSAAAAAXNSR0IArs4c6QAABGhJREFUeAHt1AEJADAMA8FtyupfVZmMh6uCcCm5M3McAQIECgKvEFJGAgQIfAGD5Q8IEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBBYSmwHshhDaqwAAAABJRU5ErkJggg==. Was data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAASwAAACWCAYAAABkW7XSAAAAAXNSR0IArs4c6QAABTlJREFUeAHt3bFtZEEQQ8HT4QLb0De0gwIY2q+BkjntEMUPYj39fD6fP/4IECBwQeDvhZAyEiBA4FfAYPkOCBA4I2CwzlQlKAECBss3QIDAGQGDdaYqQQkQMFi+AQIEzggYrDNVCUqAgMHyDRAgcEbAYJ2pSlACBAyWb4AAgTMCButMVYISIGCwfAMECJwRMFhnqhKUAAGD5RsgQOCMgME6U5WgBAgYLN8AAQJnBAzWmaoEJUDAYPkGCBA4I2CwzlQlKAECBss3QIDAGQGDdaYqQQkQ+HeR4Pv9XowtM4GUwMX/mOUXVuoTEoYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErg5/P5pAIJQ4AAgZeAX1gvGe8ECOQEDFauEoEIEHgJGKyXjHcCBHICBitXiUAECLwEDNZLxjsBAjkBg5WrRCACBF4CBusl450AgZyAwcpVIhABAi8Bg/WS8U6AQE7AYOUqEYgAgZeAwXrJeCdAICdgsHKVCESAwEvAYL1kvBMgkBMwWLlKBCJA4CVgsF4y3gkQyAkYrFwlAhEg8BIwWC8Z7wQI5AQMVq4SgQgQeAkYrJeMdwIEcgIGK1eJQAQIvAQM1kvGOwECOQGDlatEIAIEXgL/AZm6Ba13YrqjAAAAAElFTkSuQmCC.
CONSOLE MESSAGE: line 138: FAIL Should blend as if alpha is 1.0.
at (150, 75) expected: 64,64,64,255 was 0,0,0,0
CONSOLE MESSAGE: line 138: FAIL dataURL_post should be data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAASwAAACWCAYAAABkW7XSAAAAAXNSR0IArs4c6QAABGhJREFUeAHt1AEJADAMA8FtyupfVZmMh6uCcCm5M3McAQIECgKvEFJGAgQIfAGD5Q8IEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBBYSmwHshhDaqwAAAABJRU5ErkJggg==. Was data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAASwAAACWCAYAAABkW7XSAAAAAXNSR0IArs4c6QAABTlJREFUeAHt3bFtZEEQQ8HT4QLb0De0gwIY2q+BkjntEMUPYj39fD6fP/4IECBwQeDvhZAyEiBA4FfAYPkOCBA4I2CwzlQlKAECBss3QIDAGQGDdaYqQQkQMFi+AQIEzggYrDNVCUqAgMHyDRAgcEbAYJ2pSlACBAyWb4AAgTMCButMVYISIGCwfAMECJwRMFhnqhKUAAGD5RsgQOCMgME6U5WgBAgYLN8AAQJnBAzWmaoEJUDAYPkGCBA4I2CwzlQlKAECBss3QIDAGQGDdaYqQQkQ+HeR4Pv9XowtM4GUwMX/mOUXVuoTEoYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErg5/P5pAIJQ4AAgZeAX1gvGe8ECOQEDFauEoEIEHgJGKyXjHcCBHICBitXiUAECLwEDNZLxjsBAjkBg5WrRCACBF4CBusl450AgZyAwcpVIhABAi8Bg/WS8U6AQE7AYOUqEYgAgZeAwXrJeCdAICdgsHKVCESAwEvAYL1kvBMgkBMwWLlKBCJA4CVgsF4y3gkQyAkYrFwlAhEg8BIwWC8Z7wQI5AQMVq4SgQgQeAkYrJeMdwIEcgIGK1eJQAQIvAQM1kvGOwECOQGDlatEIAIEXgL/AZm6Ba13YrqjAAAAAElFTkSuQmCC.
CONSOLE MESSAGE: line 138: FAIL Should blend as if alpha is 1.0.
at (150, 75) expected: 64,64,64,255 was 0,0,0,0
CONSOLE MESSAGE: line 138: FAIL dataURL_post should be data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAASwAAACWCAYAAABkW7XSAAAAAXNSR0IArs4c6QAABGhJREFUeAHt1AEJADAMA8FtyupfVZmMh6uCcCm5M3McAQIECgKvEFJGAgQIfAGD5Q8IEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBBYSmwHshhDaqwAAAABJRU5ErkJggg==. Was data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAASwAAACWCAYAAABkW7XSAAAAAXNSR0IArs4c6QAABTlJREFUeAHt3bFtZEEQQ8HT4QLb0De0gwIY2q+BkjntEMUPYj39fD6fP/4IECBwQeDvhZAyEiBA4FfAYPkOCBA4I2CwzlQlKAECBss3QIDAGQGDdaYqQQkQMFi+AQIEzggYrDNVCUqAgMHyDRAgcEbAYJ2pSlACBAyWb4AAgTMCButMVYISIGCwfAMECJwRMFhnqhKUAAGD5RsgQOCMgME6U5WgBAgYLN8AAQJnBAzWmaoEJUDAYPkGCBA4I2CwzlQlKAECBss3QIDAGQGDdaYqQQkQ+HeR4Pv9XowtM4GUwMX/mOUXVuoTEoYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErg5/P5pAIJQ4AAgZeAX1gvGe8ECOQEDFauEoEIEHgJGKyXjHcCBHICBitXiUAECLwEDNZLxjsBAjkBg5WrRCACBF4CBusl450AgZyAwcpVIhABAi8Bg/WS8U6AQE7AYOUqEYgAgZeAwXrJeCdAICdgsHKVCESAwEvAYL1kvBMgkBMwWLlKBCJA4CVgsF4y3gkQyAkYrFwlAhEg8BIwWC8Z7wQI5AQMVq4SgQgQeAkYrJeMdwIEcgIGK1eJQAQIvAQM1kvGOwECOQGDlatEIAIEXgL/AZm6Ba13YrqjAAAAAElFTkSuQmCC.
CONSOLE MESSAGE: line 138: FAIL Should blend as if alpha is 1.0.
at (150, 75) expected: 64,64,64,255 was 0,0,0,0
CONSOLE MESSAGE: line 138: FAIL dataURL_post should be data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAASwAAACWCAYAAABkW7XSAAAAAXNSR0IArs4c6QAABGhJREFUeAHt1AEJADAMA8FtyupfVZmMh6uCcCm5M3McAQIECgKvEFJGAgQIfAGD5Q8IEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBBYSmwHshhDaqwAAAABJRU5ErkJggg==. Was data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAASwAAACWCAYAAABkW7XSAAAAAXNSR0IArs4c6QAABTlJREFUeAHt3bFtZEEQQ8HT4QLb0De0gwIY2q+BkjntEMUPYj39fD6fP/4IECBwQeDvhZAyEiBA4FfAYPkOCBA4I2CwzlQlKAECBss3QIDAGQGDdaYqQQkQMFi+AQIEzggYrDNVCUqAgMHyDRAgcEbAYJ2pSlACBAyWb4AAgTMCButMVYISIGCwfAMECJwRMFhnqhKUAAGD5RsgQOCMgME6U5WgBAgYLN8AAQJnBAzWmaoEJUDAYPkGCBA4I2CwzlQlKAECBss3QIDAGQGDdaYqQQkQ+HeR4Pv9XowtM4GUwMX/mOUXVuoTEoYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErg5/P5pAIJQ4AAgZeAX1gvGe8ECOQEDFauEoEIEHgJGKyXjHcCBHICBitXiUAECLwEDNZLxjsBAjkBg5WrRCACBF4CBusl450AgZyAwcpVIhABAi8Bg/WS8U6AQE7AYOUqEYgAgZeAwXrJeCdAICdgsHKVCESAwEvAYL1kvBMgkBMwWLlKBCJA4CVgsF4y3gkQyAkYrFwlAhEg8BIwWC8Z7wQI5AQMVq4SgQgQeAkYrJeMdwIEcgIGK1eJQAQIvAQM1kvGOwECOQGDlatEIAIEXgL/AZm6Ba13YrqjAAAAAElFTkSuQmCC.
This test runs the WebGL Test listed below in an iframe and reports PASS or FAIL.

Test: ../../resources/webgl_test_files/conformance/context/context-hidden-alpha.html
[ 1: PASS ] [object WebGLRenderingContext] is non-null.
[ 2: FAIL ] gl.getParameter(gl.ALPHA_BITS) should be 0. Was 8.
[ 3: PASS ] [object WebGLProgram] is non-null.
[ 4: FAIL ] Should blend as if alpha is 1.0. at (150, 75) expected: 64,64,64,255 was 0,0,0,0
[ 5: FAIL ] dataURL_post should be data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAASwAAACWCAYAAABkW7XSAAAAAXNSR0IArs4c6QAABGhJREFUeAHt1AEJADAMA8FtyupfVZmMh6uCcCm5M3McAQIECgKvEFJGAgQIfAGD5Q8IEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBBYSmwHshhDaqwAAAABJRU5ErkJggg==. Was data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAASwAAACWCAYAAABkW7XSAAAAAXNSR0IArs4c6QAABTlJREFUeAHt3bFtZEEQQ8HT4QLb0De0gwIY2q+BkjntEMUPYj39fD6fP/4IECBwQeDvhZAyEiBA4FfAYPkOCBA4I2CwzlQlKAECBss3QIDAGQGDdaYqQQkQMFi+AQIEzggYrDNVCUqAgMHyDRAgcEbAYJ2pSlACBAyWb4AAgTMCButMVYISIGCwfAMECJwRMFhnqhKUAAGD5RsgQOCMgME6U5WgBAgYLN8AAQJnBAzWmaoEJUDAYPkGCBA4I2CwzlQlKAECBss3QIDAGQGDdaYqQQkQ+HeR4Pv9XowtM4GUwMX/mOUXVuoTEoYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErg5/P5pAIJQ4AAgZeAX1gvGe8ECOQEDFauEoEIEHgJGKyXjHcCBHICBitXiUAECLwEDNZLxjsBAjkBg5WrRCACBF4CBusl450AgZyAwcpVIhABAi8Bg/WS8U6AQE7AYOUqEYgAgZeAwXrJeCdAICdgsHKVCESAwEvAYL1kvBMgkBMwWLlKBCJA4CVgsF4y3gkQyAkYrFwlAhEg8BIwWC8Z7wQI5AQMVq4SgQgQeAkYrJeMdwIEcgIGK1eJQAQIvAQM1kvGOwECOQGDlatEIAIEXgL/AZm6Ba13YrqjAAAAAElFTkSuQmCC.
[ 6: FAIL ] Should blend as if alpha is 1.0. at (150, 75) expected: 64,64,64,255 was 0,0,0,0
[ 7: FAIL ] dataURL_post should be data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAASwAAACWCAYAAABkW7XSAAAAAXNSR0IArs4c6QAABGhJREFUeAHt1AEJADAMA8FtyupfVZmMh6uCcCm5M3McAQIECgKvEFJGAgQIfAGD5Q8IEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBBYSmwHshhDaqwAAAABJRU5ErkJggg==. Was data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAASwAAACWCAYAAABkW7XSAAAAAXNSR0IArs4c6QAABTlJREFUeAHt3bFtZEEQQ8HT4QLb0De0gwIY2q+BkjntEMUPYj39fD6fP/4IECBwQeDvhZAyEiBA4FfAYPkOCBA4I2CwzlQlKAECBss3QIDAGQGDdaYqQQkQMFi+AQIEzggYrDNVCUqAgMHyDRAgcEbAYJ2pSlACBAyWb4AAgTMCButMVYISIGCwfAMECJwRMFhnqhKUAAGD5RsgQOCMgME6U5WgBAgYLN8AAQJnBAzWmaoEJUDAYPkGCBA4I2CwzlQlKAECBss3QIDAGQGDdaYqQQkQ+HeR4Pv9XowtM4GUwMX/mOUXVuoTEoYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErg5/P5pAIJQ4AAgZeAX1gvGe8ECOQEDFauEoEIEHgJGKyXjHcCBHICBitXiUAECLwEDNZLxjsBAjkBg5WrRCACBF4CBusl450AgZyAwcpVIhABAi8Bg/WS8U6AQE7AYOUqEYgAgZeAwXrJeCdAICdgsHKVCESAwEvAYL1kvBMgkBMwWLlKBCJA4CVgsF4y3gkQyAkYrFwlAhEg8BIwWC8Z7wQI5AQMVq4SgQgQeAkYrJeMdwIEcgIGK1eJQAQIvAQM1kvGOwECOQGDlatEIAIEXgL/AZm6Ba13YrqjAAAAAElFTkSuQmCC.
[ 8: PASS ] getError was expected value: NO_ERROR : should be no errors
[ 9: FAIL ] Should blend as if alpha is 1.0. at (150, 75) expected: 64,64,64,255 was 0,0,0,0
[ 10: FAIL ] dataURL_post should be data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAASwAAACWCAYAAABkW7XSAAAAAXNSR0IArs4c6QAABGhJREFUeAHt1AEJADAMA8FtyupfVZmMh6uCcCm5M3McAQIECgKvEFJGAgQIfAGD5Q8IEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBBYSmwHshhDaqwAAAABJRU5ErkJggg==. Was data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAASwAAACWCAYAAABkW7XSAAAAAXNSR0IArs4c6QAABTlJREFUeAHt3bFtZEEQQ8HT4QLb0De0gwIY2q+BkjntEMUPYj39fD6fP/4IECBwQeDvhZAyEiBA4FfAYPkOCBA4I2CwzlQlKAECBss3QIDAGQGDdaYqQQkQMFi+AQIEzggYrDNVCUqAgMHyDRAgcEbAYJ2pSlACBAyWb4AAgTMCButMVYISIGCwfAMECJwRMFhnqhKUAAGD5RsgQOCMgME6U5WgBAgYLN8AAQJnBAzWmaoEJUDAYPkGCBA4I2CwzlQlKAECBss3QIDAGQGDdaYqQQkQ+HeR4Pv9XowtM4GUwMX/mOUXVuoTEoYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErg5/P5pAIJQ4AAgZeAX1gvGe8ECOQEDFauEoEIEHgJGKyXjHcCBHICBitXiUAECLwEDNZLxjsBAjkBg5WrRCACBF4CBusl450AgZyAwcpVIhABAi8Bg/WS8U6AQE7AYOUqEYgAgZeAwXrJeCdAICdgsHKVCESAwEvAYL1kvBMgkBMwWLlKBCJA4CVgsF4y3gkQyAkYrFwlAhEg8BIwWC8Z7wQI5AQMVq4SgQgQeAkYrJeMdwIEcgIGK1eJQAQIvAQM1kvGOwECOQGDlatEIAIEXgL/AZm6Ba13YrqjAAAAAElFTkSuQmCC.
[ 11: FAIL ] Should blend as if alpha is 1.0. at (150, 75) expected: 64,64,64,255 was 0,0,0,0
[ 12: FAIL ] dataURL_post should be data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAASwAAACWCAYAAABkW7XSAAAAAXNSR0IArs4c6QAABGhJREFUeAHt1AEJADAMA8FtyupfVZmMh6uCcCm5M3McAQIECgKvEFJGAgQIfAGD5Q8IEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBAyWHyBAICNgsDJVCUqAgMHyAwQIZAQMVqYqQQkQMFh+gACBjIDBylQlKAECBssPECCQETBYmaoEJUDAYPkBAgQyAgYrU5WgBAgYLD9AgEBGwGBlqhKUAAGD5QcIEMgIGKxMVYISIGCw/AABAhkBg5WpSlACBBYSmwHshhDaqwAAAABJRU5ErkJggg==. Was data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAASwAAACWCAYAAABkW7XSAAAAAXNSR0IArs4c6QAABTlJREFUeAHt3bFtZEEQQ8HT4QLb0De0gwIY2q+BkjntEMUPYj39fD6fP/4IECBwQeDvhZAyEiBA4FfAYPkOCBA4I2CwzlQlKAECBss3QIDAGQGDdaYqQQkQMFi+AQIEzggYrDNVCUqAgMHyDRAgcEbAYJ2pSlACBAyWb4AAgTMCButMVYISIGCwfAMECJwRMFhnqhKUAAGD5RsgQOCMgME6U5WgBAgYLN8AAQJnBAzWmaoEJUDAYPkGCBA4I2CwzlQlKAECBss3QIDAGQGDdaYqQQkQ+HeR4Pv9XowtM4GUwMX/mOUXVuoTEoYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErAYKXqEIYAgSVgsJaOGwECKQGDlapDGAIEloDBWjpuBAikBAxWqg5hCBBYAgZr6bgRIJASMFipOoQhQGAJGKyl40aAQErg5/P5pAIJQ4AAgZeAX1gvGe8ECOQEDFauEoEIEHgJGKyXjHcCBHICBitXiUAECLwEDNZLxjsBAjkBg5WrRCACBF4CBusl450AgZyAwcpVIhABAi8Bg/WS8U6AQE7AYOUqEYgAgZeAwXrJeCdAICdgsHKVCESAwEvAYL1kvBMgkBMwWLlKBCJA4CVgsF4y3gkQyAkYrFwlAhEg8BIwWC8Z7wQI5AQMVq4SgQgQeAkYrJeMdwIEcgIGK1eJQAQIvAQM1kvGOwECOQGDlatEIAIEXgL/AZm6Ba13YrqjAAAAAElFTkSuQmCC.
[ 13: PASS ] getError was expected value: NO_ERROR : should be no errors
[ 14: PASS ] successfullyParsed is true
[ FAIL ] 9 failures reported
